requirements.txt
groq
rich
plotly
python-dotenv
